requirements.txt
beautifulsoup4
requests
openai
langchain_community
langchain
python-dotenv
duckduckgo-search


